Remove unnecessary folder creation.
Gluster creates this folder on volume creation. Whether this is new in
version 3.4 or not is unknown. If any of the hosts have this folder
prior to volume creation, then volume creation will fail.

It is not sure if this is needed when normal bricks on separate file
systems are being used. If someone wants to donate some hardware, I'm
happy to test this.