Give glusterd a chance to settle down before we create a volume.
This is usually only relevant right after initial peering if we're doing
a hot (clean) puppet build. This probably happens because puppet is
eager to run things as soon as possible, and after glusterd is
restarted, it takes a short moment for glusterd to see the other nodes
as fully peered. Ideally puppet should really only check the $onlyif
conditions *right* before it runs the command. I think that it might be
cheating here and running it in parallel slightly before... Who knows.