Improved auto repo management and version selection.
Introduced a split between repo management and version choosing.
You can now:

* Choose a package version or leave it at the default (latest).
If you choose a package version it must include the release string.
eg: in foobar-3.2.1-42.el6 the release is 42.el6
This doesn't check to see if your version is valid!

* Choose whether you want a gluster repo added automatically.
If you did specify a version, it will pick the correct repo.
This doesn't check that the repo for your os/version exists!