Add automatic generation of fs UUID's per brick filesystem.
This automatically generates UUID's for each physical filesystem, or
alternatively, you can specify one manually with the $fsuuid argument.

This will make a _big_ difference when using gluster::simple to
automatically deploy a large cluster of physical machines, since you
don't have to manually generate one uuid per device (which is time
consuming and could be a lot to do and a lot to maintain).