Add per brick defaults to gluster::simple for easier [physical] clusters.
I've had most of this patch in my head for at least a week, and I
finally got the time to implement it! If you are building a symmetrical
cluster, that has consistent device naming across all of the hosts, then
this patch is the magic that should make your life _significantly_
easier. (*cough, cough*: Ben England...)

In the corner case that some of your device have different names, you
can still use this feature in conjunction with the other parameters to
first set global defaults, and then override as needed.

If you don't specify an overriding parameter (such as $count) then the
number of elements in this array will be used as the brick count!

Please note that this patch provides the $brick_params_defaults option
which is different from the $brick_param_defaults option which will
still work, and is useful in conjunction with this option as the way to
set brick defaults across the whole cluster.

For more questions you'll be happy to see that this patch comes with
documentation and example updates.