Support distributed vagrant guest creation across a list of hosts.
If you specify a comma separated list of hosts with the --gluster-splay
argument, or with a list in the yaml file, vagrant will distribute guest
creation across the pool of hosts.

Each element can be prefixed with a <username>@ string if you need to
specify the username used to connect. By default it uses 'root'.

It is useful to pre-provision your virtual machine hosts with your
public key so that libvirt connections over ssh work without prompting.

This is particularly useful if you'd like to use vagrant to orchestrate
the building of a large gluster cluster of virtual machines across many
hosts.

Thanks to Dan Lambright for presenting me with the automation problem!