Add btrfs support.
This patch adds btrfs support to puppet-gluster. This is not (yet)
officially supported by GlusterFS but should work. It is being added
here to make it easier to test with GlusterFS and report bugs.

Special thanks to Eric Sandeen for adding --uuid support to mkfs.btrfs:

https://bugzilla.redhat.com/show_bug.cgi?id=1094857

Tracker bug for this feature is:

https://bugzilla.redhat.com/show_bug.cgi?id=1094860