Split brick_str_to_hash into separate function.
This should work around a regression introduced in a recent version of
puppet. (Perhaps 3.7.1?) Allowing custom helper functions in ruby was
apparently never "supposed" to work, and was removed silently... :(

binford2k: it was an accidental loophole got closed during code cleanup

The bug about this issue is:

https://tickets.puppetlabs.com/browse/PUP-3404

Thanks to binford2k for tracking down this bug. His URL finding skills
are impeccable!

Thanks to my puppet-gluster users for finding the issue!