Clean up / make more consistent CentOS 7 fixes.
This does some small fixes to rbarlow's original commit (merged).
Originally, I wasn't sure it was the correct patch, and I implemented
something different to try it out a different way, but in the end, I
realized that rbarlow's method was better! rbarlow, you're on your way
to becoming a puppet superstar!

This patch just makes some small changes so that the code is consistent,
and nitpitcks a bit.