Avoid repeated folder re-creation when $vrrp = false.
Since this folder won't be defined when $vrrp is false, it will get
constantly removed. It gets constantly created because the fact does an
mkdir (as it should!). As a result, ensure that it's always present.
This isn't harmful or dangerous in any way, however it adds to
unnecessary noise in your puppet logs when nothing meaningful is
happening.

Bug reported by: fubada