Add notifications to IRC.
If people dislike this change, I can revert it, but I think it's a nice
way to show people progress on the project. Also it's in #gluster-devel.

Add nobootwait to the list of options for the mount by default. This ensures that AWS EC2 instances can come up despite a potentially failed brick currently if a brick's underlying xfs filesystem has failed, the only way to recover the EC2 instance would be to perform manual repairs and ensure that the UUID matches up.

Add param to handle nobootwait or nofail depending on OS type.