Add nofail and nobootwait for respective OSes
The purpose of this PR is to allow failed fstab entries to not cause an
OS to hang on boot. This is particularly problematic on cloud providers
like AWS where there is not a quick way to manage an instance and select
a skip option prior to SSH access becoming available.