Fix a deprecation warning with stdlib 4.6.0
stdlib 4.6.0 officially deprecates the 'type' function, as it is
a reserved word for Puppet 4.0, and replaces it with a new
'type3x' function. This PR inserts a conditional to test for the
existence of a 'type3x' function and uses that. Otherwise, it will
use the 'type' function.

This change maintains the listed compatibility with stdlib versions
greater than 4.0.0.

This does not create 'future' parser or Puppet 4.0 compatibility.